refactor: restructure course form into separate components for better readability
